dbt_astro/dags/requirements.txt
apache-airflow==3.1.6
cosmos==0.1.93.3
